ml/requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
matplotlib==3.7.2
seaborn==0.12.2
jupyterlab==4.0.3
opencv-python==4.8.0.76
pillow==10.0.0
torch==2.0.1
transformers==4.31.0
flask==2.3.3
scipy==1.11.3
sentencepiece==0.1.99
protobuf==4.24.4
accelerate==0.23.0 